+++ src/main/java/com/example/Controle_Gastos/entities/Usuario.java
@@ -11,7 +11,7 @@ public class Usuario {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String nome;
 
@@ -21,16 +21,16 @@ public class Usuario {
     public Usuario() {
     }
 
-    public Usuario(Long id, String nome) {
+    public Usuario(Integer id, String nome) {
         this.id = id;
         this.nome = nome;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
